refactor: Move Translatable out of features
