+++ common/src/main/java/com/wynntils/core/consumers/functions/Function.java
@@ -5,8 +5,8 @@
 package com.wynntils.core.consumers.functions;
 
 import com.google.common.base.CaseFormat;
-import com.wynntils.core.consumers.features.Translatable;
 import com.wynntils.core.consumers.functions.arguments.FunctionArguments;
+import com.wynntils.core.persisted.Translatable;
 import java.lang.reflect.ParameterizedType;
 import java.util.List;
